Change model database(DataSiswa)

+++ adistetsa/dataprofil/models.py
@@ -11,7 +11,7 @@ DEFAULT_LENGTH = 225
 class DataSiswa(models.Model):
     NISN = models.BigIntegerField(max_length=DEFAULT_LENGTH, primary_key=True)
     NAMA = models.CharField(max_length=DEFAULT_LENGTH)
-    NIPD = models.IntegerField(max_length=DEFAULT_LENGTH)
+    NIPD = models.BigIntegerField(max_length=DEFAULT_LENGTH)
     JENIS_KELAMIN = models.CharField(
         max_length=20,
         choices=ENUM_JENIS_KELAMIN,
